Refactor file structure in components folder
