+++ src/components/common/infoCard.tsx
