+++ src/components/platforms/formatItem.tsx
