+++ src/pages/toaster.tsx
@@ -1,6 +1,6 @@
 import { createRoot } from "react-dom/client";
 import type { TwitchData, YoutubeData } from "@app-types/types";
-import Toast from "@components/toast";
+import Toast from "@components/common/toast";
 import { sizePerMinute } from "@lib/formatting";
 
 const HOST_ID = "TubeSize-Toast-Host";
